Fix linting and update to work with latest code
